Refactor receipts and block with receipts (#431)
