+++ lib/src/main/kotlin/com/swmansion/starknet/data/serializers/TransactionReceiptPolymorphicSerializer.kt
@@ -12,29 +12,13 @@ internal object TransactionReceiptPolymorphicSerializer :
         val typeElement = jsonElement.getOrElse("type") { throw SerializationException("Input element does not contain mandatory field 'type'") }
 
         val type = Json.decodeFromJsonElement(TransactionType.serializer(), typeElement)
-        val isPending = listOf("block_hash", "block_number").any { it !in jsonElement }
 
         return when (type) {
-            TransactionType.INVOKE -> when (isPending) {
-                false -> ProcessedInvokeTransactionReceipt.serializer()
-                true -> PendingInvokeTransactionReceipt.serializer()
-            }
-            TransactionType.DECLARE -> when (isPending) {
-                false -> ProcessedDeclareTransactionReceipt.serializer()
-                true -> PendingDeclareTransactionReceipt.serializer()
-            }
-            TransactionType.DEPLOY_ACCOUNT -> when (isPending) {
-                false -> ProcessedDeployAccountTransactionReceipt.serializer()
-                true -> PendingDeployAccountTransactionReceipt.serializer()
-            }
-            TransactionType.DEPLOY -> when (isPending) {
-                false -> ProcessedDeployTransactionReceipt.serializer()
-                true -> PendingDeployTransactionReceipt.serializer()
-            }
-            TransactionType.L1_HANDLER -> when (isPending) {
-                false -> ProcessedL1HandlerTransactionReceipt.serializer()
-                true -> PendingL1HandlerTransactionReceipt.serializer()
-            }
+            TransactionType.INVOKE -> InvokeTransactionReceipt.serializer()
+            TransactionType.DEPLOY_ACCOUNT -> DeployAccountTransactionReceipt.serializer()
+            TransactionType.DECLARE -> DeclareTransactionReceipt.serializer()
+            TransactionType.L1_HANDLER -> L1HandlerTransactionReceipt.serializer()
+            TransactionType.DEPLOY -> DeployTransactionReceipt.serializer()
         }
     }
 }
