+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/Responses.kt
@@ -1,11 +1,7 @@
 package com.swmansion.starknet.data.types
 
 import com.swmansion.starknet.data.serializers.HexToIntDeserializer
-import com.swmansion.starknet.data.serializers.TransactionPolymorphicSerializer
-import com.swmansion.starknet.data.serializers.TransactionReceiptPolymorphicSerializer
-import com.swmansion.starknet.data.types.transactions.Transaction
 import com.swmansion.starknet.data.types.transactions.TransactionExecutionStatus
-import com.swmansion.starknet.data.types.transactions.TransactionReceipt
 import com.swmansion.starknet.data.types.transactions.TransactionStatus
 import com.swmansion.starknet.extensions.*
 import kotlinx.serialization.ExperimentalSerializationApi
@@ -193,265 +189,6 @@ data class SyncingResponse(
     @JsonNames("highest_block_num")
     override val highestBlockNumber: Int,
 ) : Syncing()
-
-sealed interface Block {
-    val timestamp: Int
-    val sequencerAddress: Felt
-    val parentHash: Felt
-    val l1GasPrice: ResourcePrice
-    val l1DataGasPrice: ResourcePrice
-    val l1DataAvailabilityMode: L1DAMode
-    val starknetVersion: String
-}
-
-/**
- * Represents a processed block.
- *
- * Corresponds to the `BLOCK_HEADER` schema defined in the JSON-RPC spec.
- */
-sealed interface ProcessedBlock : Block {
-    val status: BlockStatus
-    val blockHash: Felt
-    val blockNumber: Int
-    val newRoot: Felt
-}
-
-/**
- * Represents a pending block.
- *
- * Corresponds to the `PENDING_BLOCK_HEADER` schema defined in the JSON-RPC spec.
- */
-sealed interface PendingBlock : Block
-
-@Serializable
-sealed class BlockWithTransactions : Block {
-    abstract val transactions: List<Transaction>
-}
-
-@Serializable
-data class ProcessedBlockWithTransactions(
-    @SerialName("status")
-    override val status: BlockStatus,
-
-    @SerialName("transactions")
-    override val transactions: List<
-        @Serializable(with = TransactionPolymorphicSerializer::class)
-        Transaction,
-        >,
-
-    @SerialName("parent_hash")
-    override val parentHash: Felt,
-
-    @SerialName("block_hash")
-    override val blockHash: Felt,
-
-    @SerialName("block_number")
-    override val blockNumber: Int,
-
-    @SerialName("new_root")
-    override val newRoot: Felt,
-
-    @SerialName("timestamp")
-    override val timestamp: Int,
-
-    @SerialName("sequencer_address")
-    override val sequencerAddress: Felt,
-
-    @SerialName("l1_gas_price")
-    override val l1GasPrice: ResourcePrice,
-
-    @SerialName("l1_data_gas_price")
-    override val l1DataGasPrice: ResourcePrice,
-
-    @SerialName("l1_da_mode")
-    override val l1DataAvailabilityMode: L1DAMode,
-
-    @SerialName("starknet_version")
-    override val starknetVersion: String,
-) : BlockWithTransactions(), ProcessedBlock
-
-@Serializable
-data class PendingBlockWithTransactions(
-    @SerialName("transactions")
-    override val transactions: List<
-        @Serializable(with = TransactionPolymorphicSerializer::class)
-        Transaction,
-        >,
-
-    @SerialName("timestamp")
-    override val timestamp: Int,
-
-    @SerialName("sequencer_address")
-    override val sequencerAddress: Felt,
-
-    @SerialName("parent_hash")
-    override val parentHash: Felt,
-
-    @SerialName("l1_gas_price")
-    override val l1GasPrice: ResourcePrice,
-
-    @SerialName("l1_data_gas_price")
-    override val l1DataGasPrice: ResourcePrice,
-
-    @SerialName("l1_da_mode")
-    override val l1DataAvailabilityMode: L1DAMode,
-
-    @SerialName("starknet_version")
-    override val starknetVersion: String,
-) : BlockWithTransactions(), PendingBlock
-
-@Serializable
-data class TransactionWithReceipt(
-    @Serializable(with = TransactionPolymorphicSerializer::class)
-    @SerialName("transaction")
-    val transaction: Transaction,
-
-    @Serializable(with = TransactionReceiptPolymorphicSerializer::class)
-    @SerialName("receipt")
-    val receipt: TransactionReceipt,
-)
-
-@Serializable
-sealed class BlockWithReceipts : Block {
-    abstract val transactionWithReceipts: List<TransactionWithReceipt>
-}
-
-@Serializable
-data class ProcessedBlockWithReceipts(
-    @SerialName("status")
-    override val status: BlockStatus,
-
-    @SerialName("transactions")
-    override val transactionWithReceipts: List<TransactionWithReceipt>,
-
-    @SerialName("block_hash")
-    override val blockHash: Felt,
-
-    @SerialName("block_number")
-    override val blockNumber: Int,
-
-    @SerialName("new_root")
-    override val newRoot: Felt,
-
-    @SerialName("timestamp")
-    override val timestamp: Int,
-
-    @SerialName("sequencer_address")
-    override val sequencerAddress: Felt,
-
-    @SerialName("parent_hash")
-    override val parentHash: Felt,
-
-    @SerialName("l1_gas_price")
-    override val l1GasPrice: ResourcePrice,
-
-    @SerialName("l1_data_gas_price")
-    override val l1DataGasPrice: ResourcePrice,
-
-    @SerialName("l1_da_mode")
-    override val l1DataAvailabilityMode: L1DAMode,
-
-    @SerialName("starknet_version")
-    override val starknetVersion: String,
-) : BlockWithReceipts(), ProcessedBlock
-
-@Serializable
-data class PendingBlockWithReceipts(
-    @SerialName("transactions")
-    override val transactionWithReceipts: List<TransactionWithReceipt>,
-
-    @SerialName("timestamp")
-    override val timestamp: Int,
-
-    @SerialName("sequencer_address")
-    override val sequencerAddress: Felt,
-
-    @SerialName("parent_hash")
-    override val parentHash: Felt,
-
-    @SerialName("l1_gas_price")
-    override val l1GasPrice: ResourcePrice,
-
-    @SerialName("l1_data_gas_price")
-    override val l1DataGasPrice: ResourcePrice,
-
-    @SerialName("l1_da_mode")
-    override val l1DataAvailabilityMode: L1DAMode,
-
-    @SerialName("starknet_version")
-    override val starknetVersion: String,
-) : BlockWithReceipts(), PendingBlock
-
-@Serializable
-sealed class BlockWithTransactionHashes : Block {
-    abstract val transactionHashes: List<Felt>
-}
-
-@Serializable
-data class ProcessedBlockWithTransactionHashes(
-    @SerialName("status")
-    override val status: BlockStatus,
-
-    @SerialName("transactions")
-    override val transactionHashes: List<Felt>,
-
-    @SerialName("block_hash")
-    override val blockHash: Felt,
-
-    @SerialName("block_number")
-    override val blockNumber: Int,
-
-    @SerialName("new_root")
-    override val newRoot: Felt,
-
-    @SerialName("timestamp")
-    override val timestamp: Int,
-
-    @SerialName("sequencer_address")
-    override val sequencerAddress: Felt,
-
-    @SerialName("parent_hash")
-    override val parentHash: Felt,
-
-    @SerialName("l1_gas_price")
-    override val l1GasPrice: ResourcePrice,
-
-    @SerialName("l1_data_gas_price")
-    override val l1DataGasPrice: ResourcePrice,
-
-    @SerialName("l1_da_mode")
-    override val l1DataAvailabilityMode: L1DAMode,
-
-    @SerialName("starknet_version")
-    override val starknetVersion: String,
-) : BlockWithTransactionHashes(), ProcessedBlock
-
-@Serializable
-data class PendingBlockWithTransactionHashes(
-    @SerialName("transactions")
-    override val transactionHashes: List<Felt>,
-
-    @SerialName("timestamp")
-    override val timestamp: Int,
-
-    @SerialName("sequencer_address")
-    override val sequencerAddress: Felt,
-
-    @SerialName("parent_hash")
-    override val parentHash: Felt,
-
-    @SerialName("l1_gas_price")
-    override val l1GasPrice: ResourcePrice,
-
-    @SerialName("l1_data_gas_price")
-    override val l1DataGasPrice: ResourcePrice,
-
-    @SerialName("l1_da_mode")
-    override val l1DataAvailabilityMode: L1DAMode,
-
-    @SerialName("starknet_version")
-    override val starknetVersion: String,
-) : BlockWithTransactionHashes(), PendingBlock
 
 @Serializable
 data class StorageEntries(
